feat(core): extract models/ into core/ kernel (Phase D)
Move all 11 model files (Zod schemas) and 5 colocated test files to
backend/src/core/models/. All models are pure type definitions with no
service dependencies. Create wildcard re-export shims at original paths.

Part of #706
