+++ backend/src/core/models/settings.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+export const SettingSchema = z.object({
+  key: z.string(),
+  value: z.string(),
+  category: z.string(),
+  updated_at: z.string(),
+});
+
+export type Setting = z.infer<typeof SettingSchema>;
